requirements.txt
flask>=2.0.0
flask-cors>=3.0.10
flask-limiter>=2.0.0
nltk>=3.7
six>=1.16.0
openai-whisper>=20231117
torch>=1.12.0
torchaudio>=0.12.0
numpy>=1.21.0
ffmpeg-python>=0.2.0
psutil>=5.9.0
pandas>=1.3.0
tqdm>=4.64.0

